+++ providers/deployable.rb
@@ -51,9 +51,11 @@ action :deploy do
 
   expected_version = "#{new_resource.version_value}#{plan_version ? ":#{plan_version}" : ''}"
 
-  directory version_dir do
-    owner node['glassfish']['user']
-    group node['glassfish']['group']
+  [archives_dir, version_dir].each do |d|
+    directory d do
+      owner node['glassfish']['user']
+      group node['glassfish']['group']
+    end
   end
 
   actual_version = nil
@@ -133,7 +135,8 @@ action :deploy do
       command << "--libraries=#{new_resource.libraries.join(',')}" unless new_resource.libraries.empty?
       command << a.target_artifact
 
-      timeout 150
+      # bash should wait for asadmin to time out first, if it doesn't because of some problem, bash should time out eventually
+      timeout node['glassfish']['asadmin']['timeout'] + 5
       user new_resource.system_user
       group new_resource.system_group
       code asadmin_command(command.join(' '))
@@ -164,9 +167,10 @@ action :undeploy do
 
     bash "asadmin_undeploy #{new_resource.component_name}" do
       unless cache_present
-        only_if "#{asadmin_command('list-applications')} #{new_resource.target}| grep -- '#{new_resource.component_name} '", :timeout => 150
+        only_if "#{asadmin_command('list-applications')} #{new_resource.target}| grep -- '#{new_resource.component_name} '", :timeout => node['glassfish']['asadmin']['timeout']
       end
-      timeout 150
+      # bash should wait for asadmin to time out first, if it doesn't because of some problem, bash should time out eventually
+      timeout node['glassfish']['asadmin']['timeout'] + 5
       user new_resource.system_user
       group new_resource.system_group
       code asadmin_command(command.join(' '))
@@ -195,8 +199,9 @@ action :disable do
   command << new_resource.component_name
 
   bash "asadmin_disable #{new_resource.component_name}" do
-    only_if "#{asadmin_command('list-applications --long')} #{new_resource.target} | grep '#{new_resource.component_name} ' | grep enabled", :timeout => 150
-    timeout 150
+    only_if "#{asadmin_command('list-applications --long')} #{new_resource.target} | grep '#{new_resource.component_name} ' | grep enabled", :timeout => node['glassfish']['asadmin']['timeout']
+    # bash should wait for asadmin to time out first, if it doesn't because of some problem, bash should time out eventually
+    timeout node['glassfish']['asadmin']['timeout'] + 5
     user new_resource.system_user
     group new_resource.system_group
     code asadmin_command(command.join(' '))
@@ -210,8 +215,9 @@ action :enable do
   command << new_resource.component_name
 
   bash "asadmin_enable #{new_resource.component_name}" do
-    not_if "#{asadmin_command('list-applications --long')} #{new_resource.target} | grep #{new_resource.component_name} | grep enabled", :timeout => 150
-    timeout 150
+    not_if "#{asadmin_command('list-applications --long')} #{new_resource.target} | grep #{new_resource.component_name} | grep enabled", :timeout => node['glassfish']['asadmin']['timeout']
+    # bash should wait for asadmin to time out first, if it doesn't because of some problem, bash should time out eventually
+    timeout node['glassfish']['asadmin']['timeout'] + 5
     user new_resource.system_user
     group new_resource.system_group
     code asadmin_command(command.join(' '))