+++ recipes/default.rb
@@ -24,13 +24,28 @@ or <code>glassfish::attribute_driven_mq</code>.
 #>
 =end
 
+# Scans Glassfish's binary for endorsed JARs and returns a list of filenames
+def gf_scan_existing_binary_endorsed_jars(install_dir)
+  jar_extensions = ['.jar']
+  gf_binary_endorsed_dir = install_dir + '/glassfish/lib/endorsed'
+  if Dir.exist?(gf_binary_endorsed_dir)
+    existing_binary_endorsed_jars = Dir.entries(gf_binary_endorsed_dir).reject { |f| File.directory?(f) || !jar_extensions.include?(File.extname(f)) }
+  else
+    existing_binary_endorsed_jars = []
+  end
+  existing_binary_endorsed_jars
+end
+
 if node['glassfish']['package_url'].nil?
-  if node['glassfish']['version'] == '4.1.152'
-    raise "The version 4.1.152 requires that node['glassfish']['variant'] be set to 'payara'" unless node['glassfish']['variant'] == 'payara'
-    node.override['glassfish']['package_url'] = 'https://s3-eu-west-1.amazonaws.com/payara.co/Payara+Downloads/payara-4.1.152.zip'
+  if node['glassfish']['version'] == '4.1.1.161'
+    raise "The version 4.1.1.161 requires that node['glassfish']['variant'] be set to 'payara'" unless node['glassfish']['variant'] == 'payara'
+    node.override['glassfish']['package_url'] = 'https://s3-eu-west-1.amazonaws.com/payara.co/Payara+Downloads/Payara+4.1.1.161/payara-4.1.1.161.zip'
   elsif node['glassfish']['version'] == '4.1.1.154'
     raise "The version 4.1.1.154 requires that node['glassfish']['variant'] be set to 'payara'" unless node['glassfish']['variant'] == 'payara'
     node.override['glassfish']['package_url'] = 'https://s3-eu-west-1.amazonaws.com/payara.co/Payara+Downloads/Payara+4.1.1.154/payara-4.1.1.154.zip'
+  elsif node['glassfish']['version'] == '4.1.152'
+    raise "The version 4.1.152 requires that node['glassfish']['variant'] be set to 'payara'" unless node['glassfish']['variant'] == 'payara'
+    node.override['glassfish']['package_url'] = 'https://s3-eu-west-1.amazonaws.com/payara.co/Payara+Downloads/payara-4.1.152.zip'
   elsif node['glassfish']['version'] == '4.1.151'
     raise "The version 4.1.151 requires that node['glassfish']['variant'] be set to 'payara'" unless node['glassfish']['variant'] == 'payara'
     node.override['glassfish']['package_url'] = 'http://s3-eu-west-1.amazonaws.com/payara.co/Payara+Downloads/payara-4.1.151.zip'
@@ -97,5 +112,35 @@ if node['glassfish']['remove_domains_dir_on_install']
     recursive true
     action :nothing
     not_if { exists_at_run_start }
+  end
+end
+
+# Install/delete endorsed JAR files into Glassfish's binary to be used thourgh the Java Endorsed mechanism.
+# see: https://docs.oracle.com/javase/7/docs/technotes/guides/standards/
+gf_binary_endorsed_dir = node['glassfish']['install_dir'] + File::Separator + 'glassfish' + File::Separator + 'lib' + File::Separator + 'endorsed'
+
+# Delete unnecessary binary endorsed jar files
+gf_scan_existing_binary_endorsed_jars(node['glassfish']['install_dir']).each do |file_name|
+  next if node['glassfish']['endorsed'] && node['glassfish']['endorsed'][file_name]
+  Chef::Log.info "Deleting binary endorsed jar file - #{file_name}"
+  file gf_binary_endorsed_dir + File::Separator + file_name do
+    action :delete
+  end
+end
+
+# Install missing binary endorsed jar files
+if node['glassfish']['endorsed']
+  node['glassfish']['endorsed'].each_pair do |file_name, value|
+    url = value['url']
+    Chef::Log.info "Installing binary endorsed jar file - #{file_name}"
+    target_file = gf_binary_endorsed_dir + File::Separator + file_name
+    remote_file target_file do
+      source url
+      mode '0600'
+      owner node['glassfish']['user']
+      group node['glassfish']['group']
+      action :create
+      not_if { ::File.exist?(target_file) }
+    end
   end
 end